tests/requirements.txt
pytest>=7.4.0
tftest>=1.8.4
pylint>=2.17.5
azure-identity>=1.14.0
azure-mgmt-resource>=23.0.1
azure-mgmt-storage>=21.1.0